leetcomp/pyproject.toml
[project]
name = "leetcomp"
version = "0.1.0"
description = "Scrapes Leetcode forum posts and aggregates compensation data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gql[requests]>=3.5.3",
    "openai>=1.99.5",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]
